Move psims imports to a separate auxiliary module for reuse
