+++ pyteomics/proforma.py
@@ -25,25 +25,13 @@ except ImportError:
 from .mass import Composition, std_aa_mass, Unimod, nist_mass, calculate_mass, std_ion_comp, mass_charge_ratio
 from .auxiliary import PyteomicsError, BasicComposition
 from .auxiliary.utils import add_metaclass
+from .auxiliary.psims import load_psimod, load_xlmod, load_gno, obo_cache, _has_psims
 
 try:
     import numpy as np
 except ImportError:
     np = None
 
-try:
-    from psims.controlled_vocabulary.controlled_vocabulary import (load_psimod, load_xlmod, load_gno, obo_cache, load_unimod)
-    _has_psims = True
-except ImportError:
-    def _needs_psims(name):
-        raise ImportError("Loading %s requires the `psims` library. To access it, please install `psims`" % name)
-
-    load_psimod = partial(_needs_psims, 'PSIMOD')
-    load_xlmod = partial(_needs_psims, 'XLMOD')
-    load_gno = partial(_needs_psims, 'GNO')
-    load_unimod = partial(_needs_psims, 'UNIMOD')
-    obo_cache = None
-    _has_psims = False
 
 _WATER_MASS = calculate_mass(formula="H2O")
